+++ packages/orq-rc/src/models/operations/updatetool.ts
@@ -1764,7 +1764,7 @@ export const UpdateToolResponseBodyCodeExecutionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M2JKTWGS24QAKB7FX5JKDCQH"),
+  _id: z.string().default("tool_01M2MM6E77GWBRCD8SE8D5K1D5"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -1959,7 +1959,7 @@ export const UpdateToolResponseBodyHTTPTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M2JKTWGREXW128GNZPN0Q6E3"),
+  _id: z.string().default("tool_01M2MM6E76HFK1JGBTZQVM07S0"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2054,7 +2054,7 @@ export const UpdateToolResponseBodyJSONSchemaTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M2JKTWGQ6K2AF3SA7NEWPMZC"),
+  _id: z.string().default("tool_01M2MM6E76XXHDZ3H6WNN2C1VC"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2157,7 +2157,7 @@ export const UpdateToolResponseBodyFunctionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M2JKTWGQ3B2HQ2CPYM750TMM"),
+  _id: z.string().default("tool_01M2MM6E75KVWZ0SPDT91998PK"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),